Changed value Types from (u)int16(_t) to uint64_t in Matrix, MatrixStorage and Dtype
Changed strided value type from std::vector to std::pair

 removed MatrixStorage::copy, because a copy c'tor will be implemented

+++ RubixMatrix/Dtype.h
@@ -15,7 +15,7 @@ namespace Rubix
 	class Dtype
 	{
 		std::string name_;
-		uint16_t physsize_; // logical size in byte of a single value of type T / Dtype.
+		uint64_t physsize_; // logical size in byte of a single value of type T / Dtype.
 
 		public:
 		Dtype()
